compression::v2: max ByteByffer::pos is uint32_t.
issue #210

+++ libdariadb/compression/v2/bytebuffer.cpp
@@ -7,7 +7,7 @@ using namespace dariadb;
 using namespace dariadb::compression::v2;
 
 ByteBuffer::ByteBuffer(const utils::Range &r) : _begin(r.begin), _end(r.end) {
-  _cap = static_cast<uint64_t>(std::distance(_begin, _end));
+  _cap = static_cast<uint32_t>(std::distance(_begin, _end));
   reset_pos();
 }
